Move deleteUser to apiCalls and update for attribute to htmlFor
