+++ src/CheckOutForm/CheckOutForm.js
@@ -1,50 +1,21 @@
 import React, { useState, useContext } from 'react';
 import { UsersContext } from '../Contexts/UsersContext';
-import { useMutation } from '@apollo/react-hooks';
-import gql from 'graphql-tag';
+import { deleteUserFromDB } from "../apiCalls/apiCalls";
 import '../CheckOutForm/CheckOutForm.css';
 
 const CheckOutForm = () => {
   const { currentUsers, setCurrentUsers } = useContext(UsersContext)
   const [search, updateSearch] = useState('');  
 
-  const deleteUserFromDB = async (id) => {
-    let options = {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    }
-
-    let queryParams = `mutation {deleteUser(id: ${id}) {user {name}}}`
-
-    let url = `https://safe-space-be.herokuapp.com/graphql?query=${queryParams}`
-
-    try {
-      let resp = await fetch(url, options);
-
-      if (!resp.ok) {
-        throw new Error('There was an error deleting the user')
-      }
-
-      let data = await resp.json();
-      console.log(data)
-      return data
-
-    } catch (error) {
-      throw error
-    }
-  }  
-
   const deleteUser = async (id) => {
-    console.log(id)
-    await deleteUserFromDB(id)
+    await deleteUserFromDB(id);
     let filteredUsers = currentUsers.original.filter(person => person.id != 3)
     setCurrentUsers({ result: filteredUsers, original: filteredUsers })
   }
 
   let namesArray = !search ? currentUsers.original : currentUsers.result 
   let namesList = namesArray.map((name, index) => {
+    console.log(name)
     return <div className="CheckOutForm_visitor button push" id="push" key={index}>
       <p>Check {name.name} Out </p>
       <button onClick={() => deleteUser(name.id)} id='delete-visitor'>X</button>
