+++ src/apiCalls/apiCalls.js
@@ -124,6 +124,32 @@ export const postNeeds = async (userId, needs) => {
     } catch (error) {
       throw error;
     }
-
   })
 };
+
+export const deleteUserFromDB = async (id) => {
+  console.log(typeof parseInt(id))
+  let options = {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    }
+  }
+
+  let queryParams = `mutation {deleteUser(id: ${id}) {user {id name}}}`
+
+  let url = `https://safe-space-be.herokuapp.com/graphql?query=${queryParams}`
+
+  try {
+    let resp = await fetch(url, options);
+
+    if (!resp.ok) {
+      throw new Error('There was an error deleting the user')
+    }
+    let data = await resp.json();
+    return data
+
+  } catch (error) {
+    throw error
+  }
+}
